requirements.txt
# Azure AI Foundry SDK
azure-ai-projects>=1.0.0
azure-ai-inference>=1.0.0
azure-identity>=1.15.0

# Observability & Logging
opencensus-ext-azure>=1.1.13
azure-monitor-opentelemetry>=1.2.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
httpx>=0.26.0
tenacity>=8.2.3
